requirements.txt
pypdf2==3.0.1
python-dateutil==2.8.2
Django==4.2
selenium==4.8.3